Add amazon link for each book to csv output

+++ server/lib/csv-helpers.js
@@ -9,25 +9,27 @@ const getCsvName = () => {
             return files
         }
     })
-    const csvFiles = files.filter(fileName => fileName.includes('.csv'));
+    const csvFiles = files.filter(fileName => fileName.includes('.csv')) || undefined;
 
     // even if more than 1 file is present, the oldest will appear first in the list and that's always the one we need to work with
     return csvFiles[0];
 }
 
 const deleteOldCsv = (csvName) => {
-    fs.unlink(`./${csvName}`, (err) => {
-        if (err) {
-            console.error(err)
-        } else {
-            console.log(`Previous export: ${csvName} was deleted`);
-        }
-    });
+    if (csvName) {
+        fs.unlink(`./${csvName}`, (err) => {
+            if (err) {
+                console.error(err)
+            } else {
+                console.log(`Previous export: ${csvName} was deleted`);
+            }
+        });
+    }
 }
 
 const exportToCsv = (booksArray, timestamp) => {
     const filename = `./${timestamp}.output.csv`;
-    const output = [['Title', 'Author', 'Goodreads Rating', 'Kindle Price', 'Hardcover Price New', 'Hardcover Price Used', 'Paperback Price New', 'Paperback Price New']]; // holds all rows of data
+    const output = [['Title', 'Author', 'Goodreads Rating', 'Kindle Price', 'Hardcover Price New', 'Hardcover Price Used', 'Paperback Price New', 'Paperback Price New', 'link']]; // holds all rows of data
 
     booksArray.forEach((book) => {
         const row = []; // a new array for each row of data
@@ -39,6 +41,7 @@ const exportToCsv = (booksArray, timestamp) => {
         row.push(`"${book.prices.hardcover.used_from}"`);
         row.push(`"${book.prices.paperback.amazon}"`);
         row.push(`"${book.prices.paperback.used_from}"`);
+        row.push(`"${book.amazon_link}"`);
 
         output.push(row.join()); // by default, join() uses a ','
     });
